Update to ServerWindow
Should only be using printText, that way you can add control characters
and they'll be parsed (TBD).
Found out that if UTF-8 doesn't know about a character, it returns a
CHARACTER REPLACEMENT character in it's place, so you should use
iso-8859-1 (or others, should make an option for this later) per line.